+++ test/test_algo_dibs.py
@@ -32,12 +32,29 @@ def test_basic(prep, tags, type, threshold, pwb):
         assert s(a) == s(b), 'line {}: "{}" != "{}"'.format(n+1, s(a), s(b))
 
 
-def test_phone_sep():
+@pytest.mark.parametrize('level', ['phone', 'syllable'])
+def test_phone_sep(level):
     text = ['hh_ih_r_;eword ',
             'dh_eh_r_;eword w_iy_;eword g_ow_;eword ']
-    sep = Separator(phone='_', syllable=None, word=';eword ')
-    model = dibs.CorpusSummary(text, separator=sep)
+
+    sep = Separator(
+        phone='_' if level == 'phone' else None,
+        syllable='_' if level == 'syllable' else None,
+        word=';eword ')
+
+    model = dibs.CorpusSummary(text, separator=sep, level=level)
     assert model.summary == {'nlines': 2, 'nwords': 4, 'nphones': 10}
+
+
+def test_bad_train(prep):
+    # cannot have a train text without word separators
+    with pytest.raises(ValueError):
+        dibs.CorpusSummary(prep)
+
+
+def test_emptyline_in_train(tags):
+    # empty lines are ignored in train file
+    dibs.CorpusSummary([''] + tags)
 
 
 def test_replicate_cdswordseg(datadir):
@@ -55,7 +72,9 @@ def test_replicate_cdswordseg(datadir):
     score = evaluate(segmented, _gold)
 
     # we obtained that score from the dibs version in CDSWordSeg
-    # (using wordseg.prepare and wordseg.evaluate in both cases)
+    # (using wordseg.prepare and wordseg.evaluate in both cases). You
+    # can replicate this result in CDSWordseg using
+    # ".../CDSwordSeg/algoComp/segment.py test/data/tagged.txt -a dibs"
     expected = {
         'type_fscore': 0.2359,
         'type_precision': 0.2084,
@@ -63,8 +82,11 @@ def test_replicate_cdswordseg(datadir):
         'token_fscore': 0.239,
         'token_precision': 0.3243,
         'token_recall': 0.1892,
-        'boundary_fscore': 0.4804,
-        'boundary_precision': 0.7161,
-        'boundary_recall': 0.3614}
+        'boundary_all_fscore': 0.6543,
+        'boundary_all_precision': 0.8377,
+        'boundary_all_recall': 0.5367,
+        'boundary_noedge_fscore': 0.4804,
+        'boundary_noedge_precision': 0.7161,
+        'boundary_noedge_recall': 0.3614}
 
     assert score == pytest.approx(expected, rel=1e-3)