+++ CHANGELOG.rst
@@ -1,3 +1,39 @@
+.. note::
+
+   Version numbers follow the `semantic versioning
+   <https://semver.org/>`_ principles.
+
+
+wordseg-0.6.2 (not yet released)
+--------------------------------
+
+* Docker image now uses python-3.6 from anaconda,
+
+* New tests to ensure replication of scores from CDSWordSeg to
+  wordseg for puddle, tp, dibs and dpseg.
+
+* In **wordseg-dpseg**:
+
+  * fixed forwarding of some arguments from Python to C++,
+  * implementation of dpseg bugfix when single char on first line of
+    a fold,
+  * use the original random number generator to replicate exactly
+    CDSWprdSeg.
+
+* In **wordseg-dibs**
+
+  * fixed bug when loading train text at syllable level (new
+    *--unit** option)
+  * safer use of train text (ensure there are word separators in
+    it).
+
+* In **wordseg-eval**, when called from bash, the scores are now
+  displayed in a fixed order. New test to ensure bash and python
+  calls to wordseg lead to identical results. See #31.
+
+* In **wordseg-stats** the scores are now displayed in a fixed order.
+
+
 wordseg-0.6.1
 -------------
 
@@ -5,7 +41,7 @@ wordseg-0.6.1
 
 * Removed dependancies to *numpy* and *pandas*.
 
-* Tests are now done on a subpart of the CHILDES corpus (was Buckeye, 
+* Tests are now done on a subpart of the CHILDES corpus (was Buckeye,
   under restrictive licence).
 
 * Simplified output in **wordseg-stats**, removed redundancy, renamed
@@ -18,10 +54,10 @@ wordseg-0.6.1
 
 * In **wordseg-ag**:
 
-  * Bugfix when compiling adaptor grammar on MacOS (removed pstream.h 
+  * Bugfix when compiling adaptor grammar on MacOS (removed pstream.h
     from AG). See #15.
-    
-  * Replaced std::tr1::unordered_{map,set} by std::unordered_{map,set}, 
+
+  * Replaced std::tr1::unordered_{map,set} by std::unordered_{map,set},
     removed useless code (custom allocator).